Renaming NavigationBreadcrumb to RouteObserverBreadcrumb to make its intended purpose more clear
